+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased][unreleased]
 
 ### Fixed
+- improve performance: remove message context menu open delay
 - improve performance: don't mark messages as seen unnecessarily when focusing window #5243
 
 <a id="1_60_0"></a>